Add files via upload
We require the fs module to work with the file system.

We specify the file paths for the source file (the one you want to read from) and the destination file (the one you want to write to).

We use fs.readFile to read the data from the source file. The 'utf8' encoding is used to read the file as a text file. If you're working with binary files, you would omit the encoding option.

Inside the callback function for fs.readFile, we check for any errors and, if there are none, we use fs.writeFile to write the data to the destination file.

In the callback function for fs.writeFile, we check for errors again and log a success message if the data was copied successfully.

Make sure to replace 'source.txt' and 'destination.txt' with the actual file paths you want to use. Also, make sure to handle errors appropriately in your production code, as this example demonstrates the basic concept.